third_party/blink/web_tests/editing/text-iterator/findString-start-search-after-selection-expected.txt

The _before_selection_ word is before the selection, so we shouldn't be able to find it if span_to_select is selected.
The _in_selection_ word is in the selection and we should always be able to find it.
The _after_selection_ word is after the selection and we should always be able to find that too.
PASS testRunner.findString("_before_selection_", []) is true
PASS testRunner.findString("_in_selection_", []) is true
PASS testRunner.findString("_after_selection_", []) is true
Selecting some text. This should make it not possible to find the _before_selection_ word without enabling wrap-around.
PASS testRunner.findString("_before_selection_", []) is false
PASS testRunner.findString("_in_selection_", []) is true
PASS testRunner.findString("_after_selection_", []) is true
PASS successfullyParsed is true

TEST COMPLETE
